fix(firmware): resolve build errors surfaced by the real ESP-IDF build
With ESP-IDF installed, check-firmware.sh now runs the actual xtensa build,
which caught pre-existing defects the x86-64 host build cannot (uint32_t is
unsigned long on the 32-bit target):

- device_controls.c: cast the 0/1 LED level to uint32_t for gpio_set_level
  (-Werror=sign-conversion).
- web_server_adapter_json.inc / web_server_login_1.inc: format uint32_t with
  "%lu" and an (unsigned long) cast instead of "%u" (-Werror=format=).
- sdkconfig.defaults: drop CONFIG_TINYUSB_HID_ENABLED, which esp_tinyusb 2.2.1
  no longer defines (HID is enabled by TINYUSB_HID_COUNT=1, already set) — it
  was an "unknown kconfig symbol" warning.
- usb_keyboard.c: esp_tinyusb 2.x owns tud_mount_cb/tud_umount_cb and reports
  attach/detach via config.event_cb; register an event handler instead of
  redefining those symbols (was a multiple-definition link error). Suspend/
  resume stay on TinyUSB's weak callbacks, which esp_tinyusb does not override.

Both projects now compile and link and the esp32s3 image builds. The USB
attach/detach handling compiles correctly but its runtime behavior still
requires on-device validation (no hardware here). clang-tidy config is
addressed separately.

+++ firmware/components/device_controls/device_controls.c
@@ -51,7 +51,7 @@ static void controls_task(void *context) {
         const uint64_t elapsed = (uint64_t)xTaskGetTickCount() * (uint64_t)portTICK_PERIOD_MS;
         const bool on = device_controls_indicator_on(get_indicator_state(), (uint32_t)elapsed);
         const int level = on ? CONFIG_APP_LED_ACTIVE_LEVEL : !CONFIG_APP_LED_ACTIVE_LEVEL;
-        if (gpio_set_level((gpio_num_t)CONFIG_APP_STATUS_LED_GPIO, level) != ESP_OK) {
+        if (gpio_set_level((gpio_num_t)CONFIG_APP_STATUS_LED_GPIO, (uint32_t)level) != ESP_OK) {
             portENTER_CRITICAL(&indicator_lock);
             indicator_state = DEVICE_INDICATOR_FATAL;
             portEXIT_CRITICAL(&indicator_lock);
